requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
blessed==1.20.0
Django==3.2.23
django-background-tasks==1.2.5
django-cgi==0.1.2
django-picklefield==3.1
django-q==1.3.9
django-windows-tools==0.2.1
jinxed==1.2.0
Pillow==10.1.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzbar==0.1.9
redis==3.5.3
six==1.16.0
sqlparse==0.4.4
types-python-dateutil==2.8.19.14
tzdata==2023.3
wcwidth==0.2.9
